requirements.txt
uvicorn==0.34.0
fastapi==0.115.11
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3